refactor : Controller/Service 구조 통일에 따라 ContestTemplate을 Contest에 통합
