+++ src/main/java/com/opus/opus/modules/contest/application/convenience/ContestConvenience.java
@@ -4,10 +4,14 @@ package com.opus.opus.modules.contest.application.convenience;
 import static com.opus.opus.modules.contest.exception.ContestExceptionType.CATEGORY_HAS_CONTEST;
 import static com.opus.opus.modules.contest.exception.ContestExceptionType.CONTEST_NAME_ALREADY_EXIST;
 import static com.opus.opus.modules.contest.exception.ContestExceptionType.NOT_FOUND_CONTEST;
+import static com.opus.opus.modules.contest.exception.ContestTeamTemplateExceptionType.NOT_FOUND_TEMPLATE;
 
 import com.opus.opus.modules.contest.domain.Contest;
+import com.opus.opus.modules.contest.domain.ContestTemplate;
 import com.opus.opus.modules.contest.domain.dao.ContestRepository;
+import com.opus.opus.modules.contest.domain.dao.ContestTemplateRepository;
 import com.opus.opus.modules.contest.exception.ContestException;
+import com.opus.opus.modules.contest.exception.ContestTeamTemplateException;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -19,6 +23,7 @@ import org.springframework.transaction.annotation.Transactional;
 public class ContestConvenience {
 
     private final ContestRepository contestRepository;
+    private final ContestTemplateRepository contestTemplateRepository;
 
     public Contest getValidateExistContest(final Long contestId) {
         return contestRepository.findById(contestId).orElseThrow(() -> new ContestException(NOT_FOUND_CONTEST));
@@ -46,5 +51,10 @@ public class ContestConvenience {
 
     public List<Contest> getCurrentContests() {
         return contestRepository.findAllByIsCurrentTrue();
+    }
+
+    public ContestTemplate getValidateExistTemplate(final Long contestId) {
+        return contestTemplateRepository.findByContestId(contestId)
+                .orElseThrow(() -> new ContestTeamTemplateException(NOT_FOUND_TEMPLATE));
     }
 }
